moneytransfer/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-countries==7.2.1
django-crispy-forms==1.12.0
django-filter==2.4.0
django-phonenumber-field==5.2.0
django-widget-tweaks==1.4.8
phonenumbers==8.12.30
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1